infra/requirements.txt
aws-cdk-lib==2.243.0
constructs==10.5.1
